+++ ios/Classes/multi_ads_sdk-Swift.h
@@ -1,8 +1,0 @@
-//
-//  multi_ads_sdk-Swift.h
-//  multi_ads_sdk
-//
-//  This file is auto-generated and should not be edited manually.
-//
-
-#import <Flutter/Flutter.h>